Extract MultiParentCasperAddBlockSpec
